Eleven findings, and the atlas found two of them independently
Two batches: six on 097bd39, five on 6809adc. The preflight gap and the
vacuous claim-support gate were also the top two findings of a codebase atlas
I built in the same window, which is some evidence they are real rather than
pedantic.

**The gate protected one call site, not the gate.** `PackLibrary.install` ran a
bounded preflight before validating; `Packs.validateFile` -- reached directly
by `verify`, by `build`, and by every `Library.open` -- did not. The validator
materializes chunk metadata and reads every text value, so a malformed file
well under any file-size ceiling exhausts the heap before it can be refused.
`PackLimits` and the preflight now live in `:pack`, where the format's ceilings
belong, and every entry point pays them.

**Zero claims scored 1.0.** A single "." survives the blank check in
`Attributions.validate`, decomposes to no claim, and the harness's empty-list
rate passed it -- so output containing no assertion at all cleared the
grounding gate and still rendered as a generated card. Vacuous is now its own
answer, because it is the case a rate cannot express.

**A replacement inherited an activation it was not allowed to keep.**
`setActive` gates broad supersessions; `--replace` copied the old pack's active
flag straight through, so a same-uid replacement was a way to make a broad
supersession live with no acknowledgement. Measured now while the old pack is
still active, and the replacement installs inactive with the impact reported.

The rest:

- A blank residual means the quotes answered everything -- spec 05 section 2.3
  says so -- and only null was handled, so a compliant generator returning ""
  had prose appended to an already-complete answer.
- Coverage checked by enumerating each row's interval: one malformed row
  reading -2e9..2e9 was two billion iterations, so the loop that was going to
  report it hung the builder instead. Interval arithmetic now.
- Table row bounds were Int while the expression's are Long, so `d6+2147483647`
  parsed as a negative range and lost its table.
- The FTS canary scanned UTF-16 chars, reading both halves of a surrogate pair
  as separators: given a supplementary-plane letter it picked a canary the
  index had never stored and refused a valid pack.
- A range form with neither bound was accepted, and can never fire.
- An unusable alias still consumed its phrase, turning two concepts into one
  conjunction -- the alias affecting the query after being dropped.
- Android cards now announce provenance before content for TalkBack, and the
  quotation-first label is part of the card's merged semantics.

567 tests; APK assembles.

+++ builder/src/main/kotlin/dev/rpghelper/builder/CorpusSpec.kt
@@ -106,8 +106,12 @@ class CorpusSpec(val root: Path, private val json: JsonObject) {
     }
 
     class RowSpec(json: JsonObject) {
-        val lo = json.number("lo").toInt()
-        val hi = json.number("hi").toInt()
+        // Long, because `DiceExpression.min`/`max` and the `table_rows` columns are. An
+        // endpoint past Int wrapped silently, so a correctly declared row for a legal
+        // expression like `d6+2147483647` -- whose first result is 2147483648 -- parsed as
+        // a negative range and took its table and roll control down with it.
+        val lo = json.number("lo").toLong()
+        val hi = json.number("hi").toLong()
         val text = json.text("text")
     }
